requirements.txt
alabaster==0.7.12
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.1
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
decouple==0.0.7
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==2.2.10
django-allauth==0.39.1
django-cors-headers==3.2.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-rest-knox==4.1.0
django-ssl-auth==2.2.0
django-summernote==0.8.11.6
djangorestframework==3.11.0
dnspython==1.16.0
docutils==0.16
email-validator==1.1.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
helpdev==0.6.10
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.0.1
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.6.2
pathtools==0.1.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psutil==5.7.0
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
PyNaCl==1.3.0
pyparsing==2.4.6
PyQt5==5.12.3
PyQt5-sip==12.7.1
PyQtWebEngine==5.12.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-decouple==3.3
python-jsonrpc-server==0.3.4
python-language-server==0.31.8
python-utils==2.4.0
python3-openid==3.1.0
pytz==2018.5
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.0
QDarkStyle==2.8
QtAwesome==0.7.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.21.0
requests-oauthlib==1.2.0
rope==0.16.0
Send2Trash==1.5.0
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==4.0.1
spyder-kernels==1.8.1
SQLAlchemy==1.3.16
sqlparse==0.2.4
stripe==2.48.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.24.2
virtualenv==20.0.7
virtualenvwrapper-win==1.2.6
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
webutils==0.9.12
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.3.1
yapf==0.29.0
zipp==3.0.0
